requirements.txt
moviepy==1.0.3
gTTS==2.3.1
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
pydub==0.25.1
